[TTS] Create separate file for module interfaces
