+++ tests/collections/tts/inference/test_text_processors.py
@@ -13,12 +13,13 @@
 # limitations under the License.
 
 import os
-import pytest
 import unittest
+
+import pytest
+from nemo_text_processing.text_normalization.normalize import Normalizer
 
 from nemo.collections.common.tokenizers.text_to_speech.tts_tokenizers import IPATokenizer
 from nemo.collections.tts.g2p.modules import IPAG2P
-from nemo_text_processing.text_normalization.normalize import Normalizer
 from nemo.collections.tts.inference.text_processors import BaseTextProcessor, IPATextTokenizer
